backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Email validation
email-validator==2.1.0

# Environment variables
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
